+++ src/main/java/roomescape/interface_adapter/Reservation/ReservationRepositoryImpl.java
@@ -8,8 +8,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.enttity.ReservationTime.ReservationTime;
-import roomescape.usecase.Reservation.Reservation;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 import roomescape.usecase.Reservation.ReservationOutput;
 import roomescape.usecase.Reservation.ReservationRepository;
