--- src/main/java/roomescape/usecase/Reservation/ReservationDto.java
@@ -1,11 +1,0 @@
-package roomescape.usecase.Reservation;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record ReservationDto(String name, LocalDate date, LocalTime time) {
-
-    public Reservation toReservation(final long id) {
-        return new Reservation(id, name, date, time);
-    }
-}
