+++ src/main/java/roomescape/usecase/Reservation/ReservationRepository.java
@@ -1,6 +1,7 @@
 package roomescape.usecase.Reservation;
 
 import java.util.List;
+import roomescape.domain.Reservation;
 
 public interface ReservationRepository {
