--- src/main/java/roomescape/usecase/Reservation/addReservationOutput.java
@@ -1,4 +1,0 @@
-package roomescape.usecase.Reservation;
-
-public class addReservationOutput {
-}
