+++ src/main/java/roomescape/usecase/ReservationTime/AddReservationTimeService.java
@@ -2,7 +2,7 @@ package roomescape.usecase.ReservationTime;
 
 import java.time.LocalTime;
 import org.springframework.stereotype.Service;
-import roomescape.enttity.ReservationTime.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 
 @Service
