+++ src/main/java/roomescape/usecase/ReservationTime/ReservationTimeOutput.java
@@ -1,7 +1,7 @@
 package roomescape.usecase.ReservationTime;
 
 import java.time.LocalTime;
-import roomescape.enttity.ReservationTime.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 public record ReservationTimeOutput(Long id, LocalTime startAt) {
     public ReservationTimeOutput(ReservationTime time) {
